test(ui): split dock suppression coverage
