+++ ui/src/components/terminal/terminal-panel.test.ts
@@ -1,65 +1,20 @@
 /* @vitest-environment jsdom */
 
-import { afterEach, beforeEach, describe, expect, it, vi, type Mock } from "vitest";
+import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 import { i18n } from "../../i18n/index.ts";
 import { createStorageMock } from "../../test-helpers/storage.ts";
 import { waitForFast } from "../../test-helpers/wait-for.ts";
 import type { TerminalGatewayClient } from "./terminal-connection.ts";
-
-type CreateOptions = {
-  parent: HTMLElement;
-  terminalOptions?: {
-    fontFamily?: string;
-    theme?: { background?: string; foreground?: string };
-  };
-  onData?: (bytes: Uint8Array) => void;
-  onResize?: (size: { columns: number; rows: number }) => void;
-};
-
-type CreateGhosttyTerminalMock = Mock<
-  (options: CreateOptions) => Promise<ReturnType<typeof createTerminalController>>
->;
-type TerminalFactory = typeof import("./terminal-runtime.ts").createIsolatedGhosttyTerminal;
+import {
+  createTerminalController,
+  defineTestTerminalPanelElement,
+  terminalOpenResult,
+  type CreateGhosttyTerminalMock,
+  type CreateOptions,
+} from "./terminal-panel.test-support.ts";
+import { OpenClawTerminalPanel } from "./terminal-panel.ts";
 
 const createGhosttyTerminalMock: CreateGhosttyTerminalMock = vi.fn();
-
-function createTerminalController(dispose: () => void = vi.fn()) {
-  const wasmTerm = {};
-  const renderer = {
-    setTheme: vi.fn(),
-    render: vi.fn(),
-  };
-  return {
-    readOnly: false,
-    terminal: {
-      cols: 100,
-      rows: 30,
-      viewportY: 0,
-      wasmTerm,
-      renderer,
-      write: vi.fn(),
-      focus: vi.fn(),
-      reset: vi.fn(),
-      paste: vi.fn(),
-    },
-    write: vi.fn(),
-    fit: vi.fn(),
-    resize: vi.fn(),
-    setReadOnly: vi.fn(),
-    attach: vi.fn(),
-    dispose,
-  };
-}
-
-function terminalOpenResult(sessionId: string) {
-  return {
-    sessionId,
-    agentId: "ops",
-    shell: "/bin/zsh",
-    cwd: "/work/ops",
-    confined: false,
-  };
-}
 
 function deferred<T>() {
   let resolve!: (value: T) => void;
@@ -71,17 +26,7 @@ function deferred<T>() {
   return { promise, resolve, reject };
 }
 
-import { OpenClawTerminalPanel } from "./terminal-panel.ts";
-
-const TERMINAL_PANEL_ELEMENT_NAME = `test-openclaw-terminal-panel-${crypto.randomUUID()}`;
-
-// The full non-isolated UI suite can import the production panel before this
-// test. Override its factory instead of relying on a module mock import order.
-class TestTerminalPanel extends OpenClawTerminalPanel {
-  override createTerminalController = createGhosttyTerminalMock as unknown as TerminalFactory;
-}
-
-customElements.define(TERMINAL_PANEL_ELEMENT_NAME, TestTerminalPanel);
+const TERMINAL_PANEL_ELEMENT_NAME = defineTestTerminalPanelElement(createGhosttyTerminalMock);
 
 async function startPanelWithPendingOpen() {
   let createOptions: CreateOptions | undefined;
@@ -136,104 +81,10 @@ describe("OpenClawTerminalPanel", () => {
     element.available = true;
     document.body.append(element);
 
-    class LazyUpgradeTerminalPanel extends TestTerminalPanel {}
-    customElements.define(tagName, LazyUpgradeTerminalPanel);
+    defineTestTerminalPanelElement(createGhosttyTerminalMock, tagName);
     const panel = element as unknown as OpenClawTerminalPanel;
     await panel.updateComplete;
     await waitForFast(() => expect(panel.terminalPanelOpen).toBe(true));
-  });
-
-  it("suppresses an open dock without overwriting its persisted preference", async () => {
-    localStorage.setItem(
-      "openclaw.terminal.panel.v1",
-      JSON.stringify({ open: true, dock: "bottom", height: 320, width: 520 }),
-    );
-    const panel = document.createElement(TERMINAL_PANEL_ELEMENT_NAME) as OpenClawTerminalPanel;
-    panel.available = true;
-    document.body.append(panel);
-    await panel.updateComplete;
-
-    expect(panel.renderRoot.querySelector(".tp")).not.toBeNull();
-    panel.suppressed = true;
-    await waitForFast(() => expect(panel.renderRoot.querySelector(".tp")).toBeNull());
-
-    expect(document.documentElement.style.getPropertyValue("--oc-terminal-reserve-bottom")).toBe(
-      "0px",
-    );
-    expect(JSON.parse(localStorage.getItem("openclaw.terminal.panel.v1") ?? "{}")).toMatchObject({
-      open: true,
-    });
-
-    panel.suppressed = false;
-    await waitForFast(() => expect(panel.renderRoot.querySelector(".tp")).not.toBeNull());
-
-    expect(panel.renderRoot.querySelector(".tp")).not.toBeNull();
-  });
-
-  it("defers availability restore until suppression ends", async () => {
-    localStorage.setItem(
-      "openclaw.terminal.panel.v1",
-      JSON.stringify({ open: true, dock: "bottom", height: 320, width: 520 }),
-    );
-    createGhosttyTerminalMock.mockResolvedValue(createTerminalController());
-    const requests: string[] = [];
-    const client: TerminalGatewayClient = {
-      forceReconnect: () => {},
-      request: async <T>(method: string) => {
-        requests.push(method);
-        return (method === "terminal.open" ? terminalOpenResult("session-1") : {}) as T;
-      },
-      addEventListener: () => () => {},
-    };
-    const panel = document.createElement(TERMINAL_PANEL_ELEMENT_NAME) as OpenClawTerminalPanel;
-    panel.client = client;
-    panel.suppressed = true;
-    document.body.append(panel);
-    await panel.updateComplete;
-
-    panel.available = true;
-    await panel.updateComplete;
-    await Promise.resolve();
-
-    expect(panel.terminalPanelOpen).toBe(false);
-    expect(requests).not.toContain("terminal.open");
-
-    panel.suppressed = false;
-    await panel.updateComplete;
-    await waitForFast(() => expect(requests).toContain("terminal.open"));
-
-    expect(panel.terminalPanelOpen).toBe(true);
-  });
-
-  it("mounts closed inside a takeover instead of booting a hidden session", async () => {
-    localStorage.setItem(
-      "openclaw.terminal.panel.v1",
-      JSON.stringify({ open: true, dock: "bottom", height: 320, width: 520 }),
-    );
-    createGhosttyTerminalMock.mockResolvedValue(createTerminalController());
-    const requests: string[] = [];
-    const client: TerminalGatewayClient = {
-      forceReconnect: () => {},
-      request: async <T>(method: string) => {
-        requests.push(method);
-        return (method === "terminal.open" ? terminalOpenResult("session-1") : {}) as T;
-      },
-      addEventListener: () => () => {},
-    };
-    const panel = document.createElement(TERMINAL_PANEL_ELEMENT_NAME) as OpenClawTerminalPanel;
-    panel.client = client;
-    panel.available = true;
-    panel.suppressed = true;
-    document.body.append(panel);
-    await panel.updateComplete;
-    await Promise.resolve();
-
-    expect(panel.terminalPanelOpen).toBe(false);
-    expect(requests).not.toContain("terminal.open");
-
-    panel.suppressed = false;
-    await panel.updateComplete;
-    await waitForFast(() => expect(requests).toContain("terminal.open"));
   });
 
   it("opens new sessions for the selected agent", async () => {
